backend/auth.py
import os
import hashlib
import hmac
from datetime import datetime, timedelta
from jose import jwt, JWTError

SECRET_KEY = os.environ.get("SECRET_KEY", "pothole-secret-key-change-in-prod-2024")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_HOURS = 24


def hash_password(password: str) -> str:
    """Simple sha256-based password hash with salt."""
    salt = os.environ.get("PASS_SALT", "ph-salt-2024")
    return hashlib.sha256(f"{salt}:{password}".encode()).hexdigest()


def verify_password(plain: str, hashed: str) -> bool:
    return hmac.compare_digest(hash_password(plain), hashed)


def create_access_token(data: dict) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(hours=ACCESS_TOKEN_EXPIRE_HOURS)
    to_encode["exp"] = expire
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_token(token: str) -> dict | None:
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except JWTError:
        return None
