backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
pillow==10.4.0
